refactor: move the view-filter editor out of the graph
`ViewFilterBuilder` lived in `components/graph/` but the graph has not used it
since saved views were removed. Its only consumer is the privacy list, through
`FileSetEditor`, so it now sits under `components/settings/` next to what
actually renders it.

The `ViewFilter*` types were misfiled the same way: defined in
`types/graph.ts`, used by the privacy filter, the Spaces membership helpers and
`lib/views.ts`, and by nothing in the graph. They move to `types/viewFilter.ts`.
`SpaceSegment` stays in `types/graph.ts` — it carries a `SegmentBy` source and
is genuinely graph-owned.

That leaves `lib/views.ts` with one remaining tie to the graph:
`buildFilterFromSegments`, a `SpaceSegment` → `ViewFilter` bridge with no
caller anywhere, plus its private `segmentToLeaf` helper. Both are removed, so
the module no longer imports from `types/graph` at all.

`components/graph/` now contains only graph code.

Note on naming: the `SpaceMembership*` helpers in `lib/views.ts` are *not*
dead despite the deprecated "Spaces" vocabulary — `PrivacyListModal` and
`dataStore` are live consumers, and the privacy list's include/exclude rules
are built on them. Renaming that vocabulary to match its actual use is worth
doing, but it is a separate change with its own blast radius.
